Update understanding to have a tail describing why we want to use HCI
Supervisor comments says it is unclear why we do designs. This helps tie together the report by saying that we want to establish whether a system fulfilling user stories is enough to solve stakeholders issues.

To this end we create a prototype and evaluate it - the result is that a system could ellivate some usability issues, but the underlaying issues regarding unused rooms remains.  Therefore, additional technological capabilites of the system is necessary.